Add ORM, templating (#5)
 - Add `pug` template engine to make generating pages easier
 - Add Sequelize ORM so we can use objects to talk to the database
 - Add a "View Addresses" page as an example of both things
 - Sugar: add `nodemon` to restart when files are changed during development.
Individual commits:

* Add MariaDB instructions

* Fix backticks

* Move mariadb to separate step

* Fix backticks again; indent

* Add download locations for MariaDB, Node.js

* Restart on file changes

Devstart will restart the server when changes are made to files, speeding up development.

* Add pug templating engine

Add

* Add partial Address page

Just a shell of a form

* Add Sequelize ORM, models

Add Sequelize, add auto-generated models based on our create.sql, add relations (may be broken).

* Remove timestamps from tables

Sequelize puts these in by default and we don't use them.

* Fix autogenerated data

This looks like a big mess, but it's replacing every '1' in the flag fields with a 0b1.

* Add relationships to objects

This lets us do user.getAddresses() and similar.  Note all the business about circular dependencies; we don't actually need to bother with it - we could just not call sync() - but it's probably for the best to know that we're being consistent with create.sql.

* Add basic "show addresses" page

So far it lists all the addresses.  It should probably just be forms to change the addresses with.