Add create/edit/delete pages for user addresses (Issue #2) (#6)
* Improve address field names

"Company name" instead of "CompanyName", etc.

* Add edit, delete, new Address buttons

They don't do anything yet though

* Move login.html to a Pug template

Not really needed, but I couldn't figure out how login.html was supposed to be hosted on my machine, and it might as well be consistent with the other templates to come.

* Add del_address endpoint

Delete an address when you click on the "Delete" button by an address.  Next up: adding/editing addresses.

* Show addresses of logged-in user

Now shows addresses for logged-in user, not hardcoded ID number

* Allow periods in usernames

Some of the users in our database have periods in the name, and the Google spec we picked says that they're OK.  This is a little more permissive than the spec, which prohibits more than one period in a row.

* Disable Sequelize sync on startup

This takes a while and spams the console with text output.  It should be moved to testing.

* Add "New Address" page

* Add validators on Address

We shouldn't be able to store empty strings in non-Null fields.

* Fix model error for User-Address relationship

* Make "New address" page work

Ta-da!  Just throws out errors to the user for now.

Adds a new address and returns the user to the address page.

* Add "Edit Address" page, functionality

You can now edit an existing address.