Change schema to single column for hash+salt
Bcrypt naturally spits out the hash and salt concatenated and there's no reason to mess with it.  Remove the PasswordHash and PasswordSalt columns and add a Password column instead.  Update data to have real hash_salt values: everyone in autogen_data has a password of "insecure", and the custom ones have some extras.